Bug 34986 - Fix the division by zero problem in OMXAudioEncoder. r=jimmy.chang
(cherry picked from commit 933a394af4d0dd015c5779758d72862f154862c2)

+++ dom/media/omx/OMXCodecWrapper.cpp
@@ -894,7 +894,7 @@ private:
   bool GetNextSubChunk(size_t& aBytesToCopy, size_t& aSamplesToCopy)
   {
     size_t bufferCapabity = GetCapacity();
-    size_t sampleBytes = mOMXAEncoder.mChannels * mOMXAEncoder.mResamplingRatio
+    double sampleBytes = mOMXAEncoder.mChannels * mOMXAEncoder.mResamplingRatio
                         * sizeof(AudioDataValue);
     if (aBytesToCopy) {
       if (aBytesToCopy > bufferCapabity) {
